Seperate RNASeq processing
