+++ oleveler/main.py
@@ -28,20 +28,12 @@ from matplotlib import ticker as mticker
 from rpy2.rinterface_lib.embedded import RRuntimeError
 import rpy2.robjects.packages as rpackages
 import rpy2.robjects as robjects
-from BCBio import GFF
-from pandas.errors import EmptyDataError
 import pandas as pd
 import numpy as np
 from collections import OrderedDict
-import re
 from hashlib import md5
-from tempfile import NamedTemporaryFile
-from threading import Thread
-from multiprocessing import Process
 from itertools import cycle
-import re
 import json
-import sys
 import pickle
 import os
 import itertools
@@ -223,321 +215,6 @@ def loadMeta(path, toRemove=[]):
     return metaDf, conditions, experiments
 
 
-def loadCountTable(tablePath, cleanId=[]):
-    """Usage:
-        loadCountTable('novogeneCounts/readcount.xls', cleanId=['gene-'])
-    Read count table and remove some SUBstring from the ID. eg. 'gene-abc' when 'gene-' in
-    cleanId list will become 'abc' in the output table.
-
-    Args:
-        tablePath (str): path to the count table
-        cleanId (list, optional): SUB string to remove from every id. Defaults to [].
-
-    Returns:
-        pd.DataFrame: count table, usually with experiments as column, genes as index.
-    """
-    logger.info(f'Load count table from file {tablePath}')
-    ct = pd.read_csv(tablePath, sep='\t', index_col=0, header=0)
-    idx = ct.index
-    for s in cleanId:
-        idx = idx.map(lambda x: x.replace(s, ''))
-    ct.index = idx
-    return ct
-
-
-def gatherCountTable(countsPath, toRemove=[]):
-    """Usage: 
-        ct = gatherCountTable('selfAlign/gene_counts')
-    Prepare count table from seperated count files produced by [featureCounts] (subread)
-
-    Args:
-        countsPath (str): path to the folder containing featureCounts outputs
-
-    Returns:
-        pd.DataFrame: count table
-    """
-    logger.info(f'Raw count table gathered from {countsPath}')
-    ct = pd.DataFrame()
-    flist = os.listdir(countsPath)
-    flist.sort()
-    for f in flist:
-        fn, ext = os.path.splitext(f)
-        if ext.lower() in ['.summary', '.log', '.gff']:
-            continue
-        elif ext == '.txt':
-            if fn in toRemove:
-                continue
-            fp = os.path.join(countsPath, f)
-            c = pd.read_csv(fp, sep='\t', comment='#',
-                            header=0, index_col=0).iloc[:, 5]
-            ct[fn] = c
-    return ct
-
-
-def _parseGff(gff, tags, getTypes=['rRNA']):
-    """Usage:
-    geneLengths, targetTypes = _parseGff('filePath', ['locus_tag'], getTypes=['rRNA'])
-
-    Do two things: 
-    1. get the length of each gene. 
-    2. get lists of target gene types listed in getTypes
-
-    Args:
-        gff (str): path to gff/gtf file
-        tags (list): a list of qualifier names that can be parsed as gene ID.
-        getTypes (list, optional): An extra list will generate with the length pdSeries.
-                                   Defaults to ['rRNA'].
-
-    Returns:
-        geneLengths (pd.Series): gene ID as index
-        targetTypes: list of lists. eg. getTypes=['rRNA','sRNA'] then 
-                     targetTypes = [['rRNA_ID1', 'rRNA_ID2',...], 
-                                    ['sRNA_ID1', 'sRNA_ID2',...]]
-    """
-
-    logger.info(f'Gathering information from annotation file {gff}, will take some time...')
-    targetTypes = [[] for t in getTypes]
-    geneLengths = pd.Series(dtype=int)
-    if isinstance(tags, str):
-        tags = [tags]
-    with open(gff, 'r') as f:
-        for rec in GFF.parse(f):
-            for feat in rec.features:
-                geneId = ''
-                if feat.type != 'gene':
-                    continue
-                for tag in tags:
-                    if tag in feat.qualifiers:
-                        geneId = feat.qualifiers[tag][0]
-                        break
-                assert geneId != '', f'Gene name not found with tag "{tags}", feature:\n{feat}'
-                geneLen = len(feat)
-                geneLengths[geneId] = geneLen
-                for i, t in enumerate(getTypes):
-                    if t in feat.qualifiers['gene_biotype']:
-                        targetTypes[i].append(geneId)
-    return geneLengths, targetTypes
-
-
-def _parseTableForGeneLength(tableFile, lengthColParsingKeys=['length'],
-                             typeCol='type', getTypes=['rRNA']):
-    """Usage:
-        geneLengths, targetTypes = \
-            _parseTableForGeneLength('tableFilePath', lengthColParsingKeys=['length'],
-                                     typeCol='gene_biotype', getTypes=['rRNA'])
-
-    Do two things: 
-    1. get the length of each gene. 
-    2. get lists of target gene types listed in getTypes
-
-    Note for ['.xls', '.tsv', '.txt'] I assume sep='\\t', for ['.csv'] I assume sep=','
-    Note I assume there is header, and geneid should be in the first columns as index
-
-    Args:
-        tableFile (str): path to table file
-        lengthColParsingKeys (list, optional): column names for gene length. Will try 
-                                               sequencially follow this list. Will stop when
-                                               one matching is found.
-                                               Defaults to ['length'].
-                                               Note: partial match is enough
-        typeCol (str, optional): column name from which you get the IDs that belong to 
-                                 target types. Defaults to 'type'. 
-                                 Note: partial match is enough
-        getTypes (list, optional): An extra list will generate with the length pdSeries.
-                                   Defaults to ['rRNA'].
-
-    Returns:
-        geneLengths (pd.Series): gene ID as index
-        targetTypes: list of lists. eg. getTypes=['rRNA','sRNA'] then 
-                     targetTypes = [['rRNA_ID1', 'rRNA_ID2',...], 
-                                    ['sRNA_ID1', 'sRNA_ID2',...]]
-    """
-    ext = os.path.splitext(tableFile)[1]
-    if ext in ['.xls', '.tsv', '.txt']:
-        sep = '\t'
-    elif ext in ['.csv']:
-        sep = ','
-    else:
-        sep = ' '
-    try:
-        tb = pd.read_csv(tableFile, sep=sep, index_col=0, header=0, comment='#')
-    except UnicodeDecodeError:  # if file extension .xls is indeed excel file
-        tb = pd.read_excel(tableFile)
-    lk = lengthColParsingKeys.copy()
-    k1 = lk.pop(0)
-    lc = [l for l in tb.columns if k1 in l]
-    while len(lc) > 1 and len(lk) > 0:
-        k = lk.pop(0)
-        lc = [l for l in lc if k in l]
-    if len(lc) > 1:
-        logger.warning(fr'''Found multiple length columns in {tableFile} with keys {lengthColParsingKeys}:
-                        {lc}
-                        Using the first one. Consider adding more key word to the selection.''')
-    lc = lc[0]
-    geneLengths = tb[lc]
-    geneLengths.name = None  # good for concating
-    targetTypes = [[] for t in getTypes]
-    typeCols = [c for c in tb.columns if typeCol in c]
-    if len(typeCols) > 0:
-        for tpc in typeCols:
-            for i, t in enumerate(getTypes):
-                targetTypes[i].extend(tb.index[tb[tpc] == t].to_list())
-    return geneLengths, targetTypes
-
-
-def _cleanCountTable(countTableDf, infoFiles,
-                     tagsForGeneName='locus_tag',
-                     lengthColParsingKeys=['length', 'gene'], typeCol='type',
-                     removerRNA=True, removeIDsubstrings=[], removeIDcontains=[]):
-    """
-    for gff files
-    infoFiles = ['a.gff', 'b.gff'] # should also support .gtf format
-    tagsForGeneName = [['tag1for_a', 'tag2for_a'], ['tag1for_b', 'tag2for_b']] # can also be str
-
-    for other tables
-    infoFiles = ['*.txt', '*.csv', '*.tsv', '*.xlsx']
-    lengthColParsingKeys=['length', 'gene'], typeCol='type', 
-
-    """
-    # Define targets
-    rrnas = []
-    geneLengths = pd.Series(dtype=int)
-
-    if isinstance(infoFiles, str):
-        infoFiles = [infoFiles]
-    if isinstance(tagsForGeneName, str):
-        tagsForGeneName = [tagsForGeneName]
-    if isinstance(lengthColParsingKeys, str):
-        lengthColParsingKeys = [lengthColParsingKeys]
-
-    gffFiles = []
-    otherFiles = []
-    for f in infoFiles:
-        ext = os.path.splitext(f)[1]
-        if ext in ['.gff', '.gtf']:
-            gffFiles.append(f)
-        else:
-            otherFiles.append(f)
-    files = gffFiles + otherFiles
-    assert len(files) >= len(tagsForGeneName)
-    tagsForGeneName += [[] for i in range(len(files)-len(tagsForGeneName))]
-
-    for i, (f, tags) in enumerate(zip(files, tagsForGeneName)):
-        if i < len(gffFiles):
-            gls, rs = _parseGff(f, tags, getTypes=['rRNA'])
-        else:
-            gls, rs = _parseTableForGeneLength(f, lengthColParsingKeys=lengthColParsingKeys,
-                                               typeCol=typeCol, getTypes=['rRNA'])
-        rrnas.extend(rs[0])
-        geneLengths = pd.concat((geneLengths, gls))
-
-    for substr in removeIDsubstrings:
-        geneLengths.index = geneLengths.index.map(lambda s: s.replace(substr, ''))
-    if removerRNA and len(rrnas) > 0:
-        logger.info(f'Found rRNA gene info in {infoFiles}, {len(rrnas)} in total, first several being:\n{rrnas[:min(5, len(rrnas))]}')
-        logger.info(f'If these ids are present in the count table, they will be removed.')
-        newct = countTableDf.loc[countTableDf.index.map(lambda x: x not in rrnas), :]
-        logger.info(f'Removed {countTableDf.shape[0] - newct.shape[0]} rRNA genes')
-    else:
-        newct = countTableDf.copy()
-    for s in removeIDcontains:
-        l = newct.shape[0]
-        newct = newct.loc[~newct.index.str.contains(s), :]
-        logger.info(f'Removed {l - newct.shape[0]} genes with "{s}" in there id')
-    return newct, geneLengths
-
-
-def _TPM(ct, geneLengths):
-    """transcripts per million
-
-    Args:
-        ct (pd.DataFrame): cleaned count table
-        geneLengths (pd.Series): lengths of every gene
-
-    Returns:
-        pd.DataFrame: calculated TPM
-    """
-    geneLengths = geneLengths[ct.index]
-    # convert to array for easy calculation
-    ctArray = ct.to_numpy()
-    glArray = geneLengths.to_numpy()
-    # normalize for gene length
-    res = ctArray/glArray[:, np.newaxis]
-    # normalize for sequencing depth
-    res = res/res.sum(axis=0)[np.newaxis, :]
-    # per million
-    res = res*1000000
-    res = pd.DataFrame(res, index=ct.index, columns=ct.columns)
-    logger.info('TPM calculated')
-    return res
-
-
-def calculateTPM(countTableDf, infoFiles,
-                 tagsForGeneName='locus_tag',
-                 lengthColParsingKeys=['length', 'gene'], typeCol='type',
-                 removerRNA=True, removeIDsubstrings=[], removeIDcontains=[]):
-    """Major function for RNA seq TPM table calculation.
-
-    # ct = prepareCountTable("selfAlign/gene_counts")
-    # calculateTPM(ct, 'selfAlign/gene_counts/GCF_000203835.1_ASM20383v1_genomic.gff', 
-    #              tagsForGeneName='locus_tag', removerRNA=True, removeIDcontains=['SCP'])
-
-    # ct = loadCountTable('novogeneCounts/readcount.xls', cleanId=['gene-'])
-    # calculateTPM(ct, 
-    #              ['novogeneCounts/gene.xls', 'novogeneCounts/sRNA_length.xls'], 
-    #              removerRNA=True, 
-    #              removeIDsubstrings=['gene-'], removeIDcontains=['sRNA', 'Novel'])
-
-    Args:
-        countTableDf (pd.DataFrame): pd.DataFrame of raw counts
-        infoFiles (str): path to annotation table or GFF/GTF files. Used in removal rRNA etc.
-        tagsForGeneName (str, optional): Feature qualifier name for parsing gene name.
-                                         Gene names should present in count table index. 
-                                         Used only when GFF/GTF files provided.
-                                         Defaults to 'locus_tag'.
-        lengthColParsingKeys (list, optional): The column names for parsing gene length,
-                                               first first, partial match. 
-                                               Defaults to ['length', 'gene'].
-        typeCol (str, optional): column name for parsing gene type (eg. to see if this gene is 
-                                 rRNA or not). Partial match. Defaults to 'type'.
-        removerRNA (bool, optional): Whether the result table include rRNA. Usually rRNA was
-                                     removed from sample before cDNA synthesis, the quantity
-                                     of removed rRNA varias and should be discarded.
-                                     Defaults to True.
-        removeIDsubstrings (list, optional): Any substrings in this list will be checked if any ID
-                                             contains them, if so, the substring will be removed from
-                                             the ID name. eg. to remove 'gene-' from 'gene-dasR' will
-                                             result in 'dasR' as the ID of this gene. 
-                                             Defaults to [].
-        removeIDcontains (list, optional): IDs that contain any substrings in this list will be
-                                           removed. Good to remove some unnecessary genes or sRNAs.
-                                           Defaults to [].
-
-    Returns:
-        pd.DataFrame: TPM calculated
-    """
-    ha = calHash(countTableDf, infoFiles,
-                 tagsForGeneName,
-                 lengthColParsingKeys, typeCol,
-                 removerRNA, removeIDsubstrings, removeIDcontains)
-    os.makedirs('dataTables', exist_ok=True)
-    tpmTablePath = f'dataTables/TPM_Table_{ha}.tsv'
-    if os.path.isfile(tpmTablePath):
-        tpmDf = pd.read_csv(tpmTablePath, sep='\t', index_col=0)
-        logger.info(f'TPM data read from {tpmTablePath}.')
-    else:
-        newct, geneLengths = _cleanCountTable(countTableDf, infoFiles,
-                                              tagsForGeneName,
-                                              lengthColParsingKeys, typeCol,
-                                              removerRNA, removeIDsubstrings, removeIDcontains)
-        tpmDf = _TPM(newct, geneLengths)
-        tpmDf.to_csv(tpmTablePath, sep='\t')
-        # Read again from file ensuring the hash stays the same if reload
-        tpmDf = pd.read_csv(tpmTablePath, sep='\t', index_col=0)
-        logger.info(f'TPM data write to table {tpmTablePath}.')
-    return tpmDf
-
-
 # ====================== DATAP ROCESSING ======================
 
 # Basic stats
@@ -640,184 +317,6 @@ def getStats(dataDf, experiments, title=''):
         logger.info(f'Write basic stats data to {f}.')
 
     return meanDf, nquantDf, varDf, stdDf, semDf
-
-
-# DESeq2 VST transformation
-def genInfoDfDESeq2(dataDf, metaDf):
-    '''
-    dataDf.column should have same elements as metaDf.Experiment
-    '''
-    infoDf = metaDf.Condition.to_frame()
-    infoDf.index = metaDf.Experiment
-    infoDf = infoDf.reindex(index=dataDf.columns)
-    tempInfoFile = NamedTemporaryFile(delete=False)
-    infoDf.index = [c.replace('-','.') for c in infoDf.index]
-    infoDf.to_csv(tempInfoFile.name, sep='\t')
-    features = infoDf.columns
-    return tempInfoFile, features
-
-
-def writeDataForDESeq2(dataDf):
-    '''
-    Will replace np.nan with zero for R import
-    A named temporary file will be used rather than writing a real data.
-    tempCleanDataFile, outForR = writeDataForR(theDf)'''
-    # fill with zeros and convert to int for DESeq2 input
-    feedToRDf = dataDf.replace(np.nan, 0).round().astype('int')
-    tempCleanDataFile = NamedTemporaryFile(delete=False)
-    feedToRDf.columns = [c.replace('-','.') for c in feedToRDf.columns]
-    feedToRDf.to_csv(tempCleanDataFile.name, sep='\t')
-    return tempCleanDataFile
-
-
-def deseq2Process(
-    dataDf,
-    metaDf,
-    ref=None,
-    deOnly=False,
-    designCol=0,
-):
-    """
-    Will return vstDf by default (deOnly = False).
-    If you want to do difference analysis, pass deOnly=True
-    """
-    pathTransformed = os.path.join('dataTables', 'transformed')
-    os.makedirs(pathTransformed, exist_ok=True)
-    # Remove columns based on metaDf
-    newCol = []
-    for e in metaDf.Experiment:
-        if e in dataDf.columns:
-            newCol.append(e)
-        else:
-            logger.info(f'{e} not in data table.')
-    dataDf = dataDf.loc[:, newCol]
-
-    # Calculate hash
-    ha = calHash(dataDf, metaDf, ref, deOnly, designCol)
-
-    logger.info(f'Current deseq2Process parameter hash = {ha}')
-
-    pathVst = os.path.join(pathTransformed, f'vst_{ha}.tsv')
-    pathVst = pathVst.replace('\\', '\\\\')
-
-    if os.path.isfile(pathVst) and not deOnly:
-        logger.info(f'Read VST transformed data from {pathVst}.')
-        vstDf = pd.read_csv(pathVst, sep='\t', header=0, index_col=0)
-        return vstDf
-
-    logger.info('####### Process data using DESeq2 #######')
-    tempDataFile = writeDataForDESeq2(dataDf)
-    tempInfoFile, features = genInfoDfDESeq2(dataDf, metaDf)
-    design = features[designCol]
-    logger.info('Fitting data using DESeq2...')
-    idxName = dataDf.index.name
-    tif = tempInfoFile.name.replace('\\', '\\\\')
-    tdf = tempDataFile.name.replace('\\', '\\\\')
-    # Clean temp file (or Windows will raise "cannot find unused tempfile name"
-    # after several loops)
-    robjects.r(
-        """
-        rm(list = ls())
-        sapply(file.path(tempdir(), list.files(tempdir())), unlink)
-        gc()
-        """
-    )
-    rpackages.importr("DESeq2")
-    # Prepare DESeq2 input
-    robjects.r(
-        f"""
-        coldata = read.csv('{tif}', sep='\t', row.names = 1, header = TRUE)
-        cts = as.matrix(read.csv('{tdf}',sep='\t', row.names = '{idxName}'))
-        """+"""
-        if (any(rownames(coldata) != colnames(cts))){
-            print("Column names not equal. Info file:")
-            print(rownames(coldata))
-            print("Data file:")
-            print(colnames(cts))
-        }
-        """
-    )
-
-    logger.info(f'Set factor for "design": "{design}"')
-    robjects.r(
-        f"""
-        coldata${design} = factor(coldata${design})
-        """
-    )
-    if isinstance(ref, type(None)):
-        ref = metaDf.loc[:, design].sort_values()[0]
-        logger.info(f'Set reference {design} (default) to {ref}')
-    else:
-        logger.info(f'Set reference {design} to {ref}')
-    robjects.r(
-        f"""
-        coldata${design} = relevel(coldata${design}, ref=as.character("{ref}"))
-        """
-    )
-
-    try:
-        robjects.r(
-            f"""
-            dds = DESeqDataSetFromMatrix(
-                countData = cts,
-                colData = coldata,
-                design = ~{design}
-            )
-            dds <- DESeq(
-                dds, test = "Wald", fitType = "parametric", sfType = "ratio",
-                betaPrior = FALSE, quiet = FALSE, useT = FALSE,
-                minmu = 0.5, parallel = FALSE
-            )
-            """
-        )
-    except RRuntimeError as e:
-        logger.info(f'Reading data with DESeq2 result error. The current environment:')
-        logger.info('column data')
-        logger.info(robjects.r('coldata'))
-        logger.info('data table')
-        logger.info(robjects.r('head(cts)'))
-        raise(e)
-
-    tempInfoFile.close()
-    tempDataFile.close()
-    os.remove(tempInfoFile.name)
-    os.remove(tempDataFile.name)
-    logger.info('END DESeq2 data init with current design:')
-    if deOnly:
-        return design
-    else:
-        logger.info('Performing VST')
-
-        # Calculate nsub for vst
-        robjects.r(
-            """
-            nsub <- 1000
-            baseMean <- rowMeans(counts(dds, normalized=TRUE))
-            max_nsub <- sum(baseMean > 5) 
-            if (max_nsub < 1000) {
-                nsub <- max_nsub
-                print("Less than 1000 rows with mean normalized count > 5.")
-                print(nsub)
-            }
-            """
-        )
-
-        robjects.r(
-            f"""
-            vsd = vst(dds, nsub=nsub, blind=FALSE)
-            write.table(assay(vsd), file='{pathVst}', sep='\t', col.names=NA)
-            rm(vsd)
-            gc()
-            """
-        )
-        logger.info(pathVst)
-        logger.info('####### END Processing data using DESeq2 #######')
-        vstDf = pd.read_csv(pathVst, sep='\t', header=0, index_col=0)
-        # sort columns for the result.
-        vstDf = vstDf.loc[:, sorted(vstDf.columns)]
-        vstDf.to_csv(pathVst, sep='\t')
-        vstDf = pd.read_csv(pathVst, sep='\t', header=0, index_col=0)
-        return vstDf
 
 
 # MSstats proposition
@@ -1040,144 +539,6 @@ def msstatsComp(comparisons, mqDataPath, annotationPath, compResultFile, toRemov
     return compResultFile
 
 
-def _deseq2Comp(a, b, name, extra, subResFileName):
-    # First create environment for use in seperate threads or processes
-    rstr = f"""
-        #print(resultsNames(dds))
-        res <- {a}(dds, {b}="{name}", parallel=TRUE{extra})
-        
-        write.table(as.data.frame(res), col.names=NA, sep='\t',
-                    file="{subResFileName}")
-        #print("{subResFileName}")
-        """
-    # Do not directly use f string in r() function here, for maximal compatibility with
-    # Thread() in Windows
-    robjects.r(rstr)
-
-
-def deseq2Comp(comparisons, countTable, annotationPath, compResultFile, lfcShrink=True, timeout=100):
-    """Warning: Do not make too many comparisions in the comparisons file, it will cause R to consume too
-    much memory and stop.
-
-    Args:
-        comparisons ([type]): [description]
-        countTable ([type]): [description]
-        annotationPath ([type]): [description]
-        compResultFile ([type]): [description]
-        lfcShrink (bool, optional): [description]. Defaults to True.
-        timeout (int, optional): Note this will cause error if the processing in R indeed cost much time. Defaults to 300.
-
-    Returns:
-        [type]: [description]
-    """
-    metaDf, _, _ = loadMeta(annotationPath)
-    allCompResults = OrderedDict()
-    previousCtr = ''
-    previousDesign = ''
-    # DESeq2 dds is control condition sensitive.
-    # Sort the comparisons based on the control conditions to reduce time consumption
-    # in creating dds in DESeq2 with same control.
-    sortedComps = sorted(comparisons, key=lambda x: comparisons[x]['ctr'])
-    nDeseqRuns = len(set(comparisons[c]['ctr'] for c in comparisons))
-    if nDeseqRuns > 4:
-        logger.warning(rf'''There are more than 5 different controls to run, note the program may crash.
-        please restart python kernel (thus you also killed R kernel) and run from the first cell.
-        If that don't work, then you need to reduce the number of controls from the comparisons.
-        Or you can use a better computer with larget memory to try.''')
-    for c in sortedComps:
-        exp = comparisons[c]['exp']
-        ctr = comparisons[c]['ctr']
-        logger.info(f'Calculate comparisons {exp} vs. {ctr}.')
-        if ctr == previousCtr:
-            design = previousDesign
-        else:
-            design = deseq2Process(countTable, metaDf, ref=ctr, deOnly=True)
-            previousDesign = design
-            previousCtr = ctr
-        name = f'{design}_{exp}_vs_{ctr}'
-        subResFile = NamedTemporaryFile(delete=False)
-        if lfcShrink:
-            a = 'lfcShrink'
-            b = 'coef'
-            extra = ', type="apeglm"'
-            logger.info('Using lfcShrink of the log2fc, ' +
-                        'note the column "lfcSE" in shrinked data is actually "posterior SD"')
-        else:
-            a = 'results'
-            b = 'name'
-            extra = ', alpha=0.05'
-        # Code below within this loop is to prevent R from lose respond.
-        # Sometime (more likely to happen with lfcShrink) the calculation is done but python
-        # does not get the signal to proceed.
-        # Once a time out is added, then the possiblity that R cannot finish on time have to
-        # be considered. The solution is to increase given time in a while loop.
-        # rpy works by running a Python process and an R process in parallel,
-        # and exchange information between them.
-        # It does not take into account that R calls are called in parallel using multiprocess.
-        # So in practice, each of the python processes connects to the same R process.
-        maxRetryTime = 30*60  # 30 min
-        i = 0
-        t = timeout
-        #fn = subResFile.name
-        #print('temp file name ', fn)
-        subResFile.close()
-        #print('temp file name after close ', subResFile.name)
-        
-        while os.path.getsize(subResFile.name) == 0:
-            # print(os.path.getsize(subResFile.name))
-            if i > 0:
-                logger.info(f'Failed to calculate comparison within {int(t/60)} min, retry.')
-            srf = subResFile.name.replace('\\', '\\\\')
-
-            # Add time out to function. Trust me with Thread or Process!!!
-            # _deseq2Comp() is an function calling R code by rpy2, needs to communicate with the
-            # only r kernel managed by rpy2. Add time out is just to retry when r kernel do not
-            # respond.
-            # In *nix system with Thread(), the died kernel stays no respond, so Process() needed.
-            # In Win, Process() will lose the contact with rpy2 kernel, only Thread() is safe and
-            # do not have the problem of no responding r kernel (or do not have the problem like
-            # above).
-            # Maybe experiment with spawn and fork?
-            if sys.platform.startswith("win") or sys.platform == 'darwin':
-                th = Thread(target=_deseq2Comp, args=(a, b, name, extra, srf), daemon=True)
-                th.start()
-                th.join(t)
-            else:
-                p = Process(target=_deseq2Comp, args=(a, b, name, extra, srf), daemon=True)
-                p.start()
-                p.join(timeout=t)
-            t += timeout * max(0, i-1)  # retry once with the same time out, then increase this time.
-            if t > maxRetryTime:
-                break
-            i += 1
-        try:
-            subResDf = pd.read_csv(subResFile.name, sep='\t', index_col=0)
-            allCompResults[c] = subResDf
-        except EmptyDataError:
-            logger.warning('Failed calculating comparison with ' +
-                           f'{i} trials, max run time per trial is {maxRetryTime/60} min. ' +
-                           f'The comparison {exp} vs {ctr} with name {c} ignored.')
-        os.remove(subResFile.name)
-    # Write to compResultFile
-    # Generate a dataframe that store all reulsts in one table like MSstats
-    allCompDf = pd.DataFrame()
-    for c in comparisons:
-        subResDf = allCompResults[c].copy()
-        cols = ['Label'] + subResDf.columns.to_list()
-        subResDf["Label"] = c
-        subResDf.index.name = 'ID'
-        subResDf = subResDf[cols]
-        # Move  index to columns
-        subResDf = subResDf.reset_index()
-        allCompDf = pd.concat([allCompDf, subResDf], axis=0, ignore_index=True)
-    allCompDf = allCompDf.set_index('ID', drop=True)
-    targetColsOri = ['Label', 'log2FoldChange', 'lfcSE', 'pvalue', 'padj',       'baseMean']
-    targetCols = ['Label', 'log2FC',         'SE',    'pvalue', 'adj.pvalue', 'baseMean']
-    allCompDf = allCompDf.loc[:, targetColsOri]
-    allCompDf.columns = targetCols
-    allCompDf.to_csv(compResultFile, sep='\t')
-    return compResultFile
-
 
 def genComparisonResults(compResultFile, comparisons):
     """
@@ -1287,61 +648,6 @@ def makeCompMatrixMsstats(compExcel, mqDataPath, annotationPath, toRemove=[]):
     logger.info(f'Done calculation, dump data in {compResultFile}\n')
     clearlogfiles()
     return allCompResults, comparisons
-
-
-def makeCompMatrixDeseq2(compExcel, countTable, annotationPath, shrink=None):
-    '''
-    shrink only affects lfc values, not pvalues
-
-    # Convert to data tables for each comparison.
-    #
-    # > The result of the test for diffrential abundance is a table with
-    # > columns Protein, Label (of the comparison), log2 fold change (log2FC), 
-    # > standard error of the log2 fold change (SE), test statistic of the 
-    # > Student test (Tvalue), degree of freedom of the Student test (DF), 
-    # > raw p-values (pvalue), p-values adjusted among all the proteins in 
-    # > the specific comparison using the approach by Benjamini and Hochberg 
-    # > (adj.pvalue). The cutoff of the adjusted p-value corresponds to the 
-    # > cutoff of the False Discovery Rate (Benjamini and Hochberg 1955). 
-    # > The positive values of log2FC for Label=C2-C1 indicate evidence in 
-    # > favor of C2 > C1 (i.e. proteins upregulated in C2), while the negative 
-    # > values indicate evidence in favor of C2 < C1 (i.e. proteins 
-    # > downregulated in C2), as compared to C1. The same model can be used 
-    # > to perform several comparisons of conditions simultaneously in the 
-    # > same protein.
-    # >
-    # > **NOTE** `issue` column shows if there is any issue for inference in 
-    # > corresponding protein and comparison, for example, OneConditionMissing 
-    # > or CompleteMissing. If one of condition for compariosn is completely 
-    # > missing, it would flag with OneConditionMissiong with adj.pvalue=0 and 
-    # > log2FC=Inf or -Inf even though pvalue=NA. For example, if you want to 
-    # > compare ‘Condition A - Condition B’, but condition B has complete missing, 
-    # > log2FC=Inf and adj.pvalue=0. SE, Tvalue, and pvalue will be NA. if you 
-    # > want to compare ‘Condition A - Condition B’, but condition A has complete 
-    # > missing, then log2FC=-Inf and adj.pvalue=0. But, please be careful for 
-    # > using this log2FC and adj.pvalue.
-    '''
-    # Process input see if anything has changed.
-    # If not, read from existing result if have.
-    if shrink == 'lfcShrink':
-        compResultFileBase = 'dataTables/transformed/deseq2_comparisonResult_lfcShrink.tsv'
-    elif isinstance(shrink, type(None)):
-        compResultFileBase = 'dataTables/transformed/deseq2_comparisonResult.tsv'
-    else:
-        raise ValueError(f'shrink should be in [None, "lfcShrink"], but now it is {shrink}')
-
-    allCompResults, comparisons, compResultFile = \
-        _checkExistingCompResult(compExcel, countTable, compResultFileBase)
-    if not isinstance(allCompResults, type(None)):
-        return allCompResults, comparisons
-
-    logger.info(f'Parse {compExcel}, generate comparison matrix using DESeq2.')
-    compResultFile = deseq2Comp(comparisons, countTable, annotationPath, compResultFile,
-                                lfcShrink=(True if shrink == 'lfcShrink' else False))
-    allCompResults = genComparisonResults(compResultFile, comparisons)
-    logger.info(f'Done calculation, dump data in {compResultFile}\n')
-    return allCompResults, comparisons
-
 
 def getSig(compDf, tFc=1.5, tPv=0.05):
     """return compDf with only significant ids
